+++ standalone-metastore/metastore-server/src/main/java/org/apache/hadoop/hive/common/histogram/KllHistogramEstimator.java
@@ -15,9 +15,10 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.hadoop.hive.ql.udf.datasketches.kll;
+package org.apache.hadoop.hive.common.histogram;
 
 import org.apache.datasketches.kll.KllFloatsSketch;
+import org.apache.hadoop.hive.common.histogram.kll.KllUtils;
 import org.apache.hadoop.hive.common.type.HiveDecimal;
 import org.apache.hadoop.hive.ql.util.JavaDataModel;
 
@@ -27,6 +28,10 @@ import java.io.IOException;
 public class KllHistogramEstimator {
 
   private final KllFloatsSketch kll;
+
+  KllHistogramEstimator() {
+    this.kll = new KllFloatsSketch();
+  }
 
   KllHistogramEstimator(int k) {
     this.kll = new KllFloatsSketch(k);
@@ -68,11 +73,15 @@ public class KllHistogramEstimator {
     return KllUtils.lengthFor(model, kll);
   }
 
-  public KllFloatsSketch getSketch() {
-    return kll;
+  public boolean canMerge(KllHistogramEstimator o) {
+    return o != null && this.kll.getK() == o.kll.getK();
   }
 
   public int getK() {
     return kll.getK();
+  }
+
+  public KllFloatsSketch getSketch() {
+    return kll;
   }
 }
